Split DataObject object definitions, no functional changes
