+++ src/System.Windows.Forms/src/System/Windows/Forms/DataObject.cs
@@ -11,11 +11,9 @@ using System.Diagnostics;
 using System.Drawing;
 using System.Globalization;
 using System.IO;
-using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Runtime.InteropServices.ComTypes;
 using System.Runtime.Serialization;
-using System.Runtime.Serialization.Formatters;
 using System.Runtime.Serialization.Formatters.Binary;
 using System.Text;
 using IComDataObject = System.Runtime.InteropServices.ComTypes.IDataObject;
@@ -27,7 +25,7 @@ namespace System.Windows.Forms
     ///  Implements a basic data transfer mechanism.
     /// </summary>
     [ClassInterface(ClassInterfaceType.None)]
-    public class DataObject : IDataObject, IComDataObject
+    public partial class DataObject : IDataObject, IComDataObject
     {
         private static readonly string CF_DEPRECATED_FILENAME = "FileName";
         private static readonly string CF_DEPRECATED_FILENAMEW = "FileNameW";
@@ -1089,1123 +1087,6 @@ namespace System.Windows.Forms
             Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "Set data: " + data?.ToString() ?? "(null)");
             Debug.Assert(innerData != null, "You must have an innerData on all DataObjects");
             innerData.SetData(data);
-        }
-
-        /// <summary>
-        ///  Part of IComDataObject, used to interop with OLE.
-        /// </summary>
-        private class FormatEnumerator : IEnumFORMATETC
-        {
-            internal IDataObject parent = null;
-            internal ArrayList formats = new ArrayList();
-            internal int current = 0;
-
-            public FormatEnumerator(IDataObject parent) : this(parent, parent.GetFormats())
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "FormatEnumerator: Constructed: " + parent.ToString());
-            }
-
-            public FormatEnumerator(IDataObject parent, FORMATETC[] formats)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "FormatEnumerator: Constructed: " + parent.ToString() + ", FORMATETC[]" + formats.Length.ToString(CultureInfo.InvariantCulture));
-                this.formats.Clear();
-                this.parent = parent;
-                current = 0;
-                if (formats != null)
-                {
-                    if (parent is DataObject dataObject && dataObject.RestrictedFormats)
-                    {
-                        if (!Clipboard.IsFormatValid(formats))
-                        {
-                            throw new Security.SecurityException(SR.ClipboardSecurityException);
-                        }
-                    }
-
-                    for (int i = 0; i < formats.Length; i++)
-                    {
-                        FORMATETC currentFormat = formats[i];
-
-                        FORMATETC temp = new FORMATETC
-                        {
-                            cfFormat = currentFormat.cfFormat,
-                            dwAspect = currentFormat.dwAspect,
-                            ptd = currentFormat.ptd,
-                            lindex = currentFormat.lindex,
-                            tymed = currentFormat.tymed
-                        };
-                        this.formats.Add(temp);
-                    }
-                }
-            }
-
-            public FormatEnumerator(IDataObject parent, string[] formats)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "FormatEnumerator: Constructed: " + parent.ToString() + ", string[]" + formats.Length.ToString(CultureInfo.InvariantCulture));
-
-                this.parent = parent;
-                this.formats.Clear();
-
-                if (formats != null)
-                {
-                    if (parent is DataObject dataObject && dataObject.RestrictedFormats)
-                    {
-                        if (!Clipboard.IsFormatValid(formats))
-                        {
-                            throw new Security.SecurityException(SR.ClipboardSecurityException);
-                        }
-                    }
-
-                    for (int i = 0; i < formats.Length; i++)
-                    {
-                        string format = formats[i];
-                        FORMATETC temp = new FORMATETC
-                        {
-                            cfFormat = unchecked((short)(ushort)(DataFormats.GetFormat(format).Id)),
-                            dwAspect = DVASPECT.DVASPECT_CONTENT,
-                            ptd = IntPtr.Zero,
-                            lindex = -1
-                        };
-
-                        if (format.Equals(DataFormats.Bitmap))
-                        {
-                            temp.tymed = TYMED.TYMED_GDI;
-                        }
-                        else if (format.Equals(DataFormats.EnhancedMetafile))
-                        {
-                            temp.tymed = TYMED.TYMED_ENHMF;
-                        }
-                        else if (format.Equals(DataFormats.Text)
-                                 || format.Equals(DataFormats.UnicodeText)
-                                 || format.Equals(DataFormats.StringFormat)
-                                 || format.Equals(DataFormats.Rtf)
-                                 || format.Equals(DataFormats.CommaSeparatedValue)
-                                 || format.Equals(DataFormats.FileDrop)
-                                 || format.Equals(CF_DEPRECATED_FILENAME)
-                                 || format.Equals(CF_DEPRECATED_FILENAMEW))
-                        {
-                            temp.tymed = TYMED.TYMED_HGLOBAL;
-                        }
-                        else
-                        {
-                            temp.tymed = TYMED.TYMED_HGLOBAL;
-                        }
-
-                        if (temp.tymed != 0)
-                        {
-                            this.formats.Add(temp);
-                        }
-                    }
-                }
-            }
-
-            public int Next(int celt, FORMATETC[] rgelt, int[] pceltFetched)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "FormatEnumerator: Next");
-                if (this.current < formats.Count && celt > 0)
-                {
-                    FORMATETC current = (FORMATETC)formats[this.current];
-                    rgelt[0].cfFormat = current.cfFormat;
-                    rgelt[0].tymed = current.tymed;
-                    rgelt[0].dwAspect = DVASPECT.DVASPECT_CONTENT;
-                    rgelt[0].ptd = IntPtr.Zero;
-                    rgelt[0].lindex = -1;
-
-                    if (pceltFetched != null)
-                    {
-                        pceltFetched[0] = 1;
-                    }
-                    this.current++;
-                }
-                else
-                {
-                    if (pceltFetched != null)
-                    {
-                        pceltFetched[0] = 0;
-                    }
-                    return (int)HRESULT.S_FALSE;
-                }
-                return (int)HRESULT.S_OK;
-            }
-
-            public int Skip(int celt)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "FormatEnumerator: Skip");
-                if (current + celt >= formats.Count)
-                {
-                    return (int)HRESULT.S_FALSE;
-                }
-                current += celt;
-                return (int)HRESULT.S_OK;
-            }
-
-            public int Reset()
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "FormatEnumerator: Reset");
-                current = 0;
-                return (int)HRESULT.S_OK;
-            }
-
-            public void Clone(out IEnumFORMATETC ppenum)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "FormatEnumerator: Clone");
-                FORMATETC[] temp = new FORMATETC[formats.Count];
-                formats.CopyTo(temp, 0);
-                ppenum = new FormatEnumerator(parent, temp);
-            }
-        }
-
-        /// <summary>
-        ///  OLE Converter.  This class embodies the nastiness required to convert from our
-        ///  managed types to standard OLE clipboard formats.
-        /// </summary>
-        private class OleConverter : IDataObject
-        {
-            internal IComDataObject innerData;
-
-            public OleConverter(IComDataObject data)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "OleConverter: Constructed OleConverter");
-                innerData = data;
-            }
-
-            /// <summary>
-            ///  Returns the data Object we are wrapping
-            /// </summary>
-            public IComDataObject OleDataObject
-            {
-                get
-                {
-                    return innerData;
-                }
-            }
-
-            /// <summary>
-            ///  Uses IStream and retrieves the specified format from the bound IComDataObject.
-            /// </summary>
-            private unsafe object GetDataFromOleIStream(string format)
-            {
-                FORMATETC formatetc = new FORMATETC();
-                STGMEDIUM medium = new STGMEDIUM();
-
-                formatetc.cfFormat = unchecked((short)(ushort)(DataFormats.GetFormat(format).Id));
-                formatetc.dwAspect = DVASPECT.DVASPECT_CONTENT;
-                formatetc.lindex = -1;
-                formatetc.tymed = TYMED.TYMED_ISTREAM;
-
-                medium.tymed = TYMED.TYMED_ISTREAM;
-
-                // Limit the # of exceptions we may throw below.
-                if ((int)HRESULT.S_OK != QueryGetDataUnsafe(ref formatetc))
-                {
-                    return null;
-                }
-
-                try
-                {
-                    innerData.GetData(ref formatetc, out medium);
-                }
-                catch
-                {
-                    return null;
-                }
-
-                if (medium.unionmember != IntPtr.Zero)
-                {
-                    Ole32.IStream pStream = (Ole32.IStream)Marshal.GetObjectForIUnknown(medium.unionmember);
-                    Marshal.Release(medium.unionmember);
-                    pStream.Stat(out Ole32.STATSTG sstg, Ole32.STATFLAG.STATFLAG_DEFAULT);
-
-                    IntPtr hglobal = Kernel32.GlobalAlloc(
-                        Kernel32.GMEM.MOVEABLE | Kernel32.GMEM.DDESHARE | Kernel32.GMEM.ZEROINIT,
-                        (uint)sstg.cbSize);
-                    IntPtr ptr = Kernel32.GlobalLock(hglobal);
-                    pStream.Read((byte*)ptr, (uint)sstg.cbSize, null);
-                    Kernel32.GlobalUnlock(hglobal);
-
-                    return GetDataFromHGLOBAL(format, hglobal);
-                }
-
-                return null;
-            }
-
-            /// <summary>
-            ///  Retrieves the specified form from the specified hglobal.
-            /// </summary>
-            private object GetDataFromHGLOBAL(string format, IntPtr hglobal)
-            {
-                object data = null;
-
-                if (hglobal != IntPtr.Zero)
-                {
-                    // Convert from OLE to IW objects
-                    if (format.Equals(DataFormats.Text)
-                        || format.Equals(DataFormats.Rtf)
-                        || format.Equals(DataFormats.OemText))
-                    {
-                        data = ReadStringFromHandle(hglobal, false);
-                    }
-                    else if (format.Equals(DataFormats.Html))
-                    {
-                        data = ReadHtmlFromHandle(hglobal);
-                    }
-
-                    else if (format.Equals(DataFormats.UnicodeText))
-                    {
-                        data = ReadStringFromHandle(hglobal, true);
-                    }
-                    else if (format.Equals(DataFormats.FileDrop))
-                    {
-                        data = ReadFileListFromHandle(hglobal);
-                    }
-                    else if (format.Equals(CF_DEPRECATED_FILENAME))
-                    {
-                        data = new string[] { ReadStringFromHandle(hglobal, false) };
-                    }
-                    else if (format.Equals(CF_DEPRECATED_FILENAMEW))
-                    {
-                        data = new string[] { ReadStringFromHandle(hglobal, true) };
-                    }
-                    else
-                    {
-                        data = ReadObjectFromHandle(hglobal, DataObject.RestrictDeserializationToSafeTypes(format));
-                    }
-
-                    Kernel32.GlobalFree(hglobal);
-                }
-
-                return data;
-            }
-
-            /// <summary>
-            ///  Uses HGLOBALs and retrieves the specified format from the bound IComDatabject.
-            /// </summary>
-            private object GetDataFromOleHGLOBAL(string format, out bool done)
-            {
-                done = false;
-                Debug.Assert(innerData != null, "You must have an innerData on all DataObjects");
-
-                FORMATETC formatetc = new FORMATETC();
-                STGMEDIUM medium = new STGMEDIUM();
-
-                formatetc.cfFormat = unchecked((short)(ushort)(DataFormats.GetFormat(format).Id));
-                formatetc.dwAspect = DVASPECT.DVASPECT_CONTENT;
-                formatetc.lindex = -1;
-                formatetc.tymed = TYMED.TYMED_HGLOBAL;
-
-                medium.tymed = TYMED.TYMED_HGLOBAL;
-
-                object data = null;
-
-                if ((int)HRESULT.S_OK == QueryGetDataUnsafe(ref formatetc))
-                {
-                    try
-                    {
-                        innerData.GetData(ref formatetc, out medium);
-
-                        if (medium.unionmember != IntPtr.Zero)
-                        {
-                            data = GetDataFromHGLOBAL(format, medium.unionmember);
-                        }
-                    }
-                    catch (RestrictedTypeDeserializationException)
-                    {
-                        done = true;
-                    }
-                    catch
-                    {
-                    }
-                }
-                return data;
-            }
-
-            /// <summary>
-            ///  Retrieves the specified format data from the bound IComDataObject, from
-            ///  other sources that IStream and HGLOBAL... this is really just a place
-            ///  to put the "special" formats like BITMAP, ENHMF, etc.
-            /// </summary>
-            private object GetDataFromOleOther(string format)
-            {
-                Debug.Assert(innerData != null, "You must have an innerData on all DataObjects");
-
-                FORMATETC formatetc = new FORMATETC();
-                STGMEDIUM medium = new STGMEDIUM();
-
-                TYMED tymed = (TYMED)0;
-
-                if (format.Equals(DataFormats.Bitmap))
-                {
-                    tymed = TYMED.TYMED_GDI;
-                }
-                else if (format.Equals(DataFormats.EnhancedMetafile))
-                {
-                    tymed = TYMED.TYMED_ENHMF;
-                }
-
-                if (tymed == (TYMED)0)
-                {
-                    return null;
-                }
-
-                formatetc.cfFormat = unchecked((short)(ushort)(DataFormats.GetFormat(format).Id));
-                formatetc.dwAspect = DVASPECT.DVASPECT_CONTENT;
-                formatetc.lindex = -1;
-                formatetc.tymed = tymed;
-                medium.tymed = tymed;
-
-                object data = null;
-                if ((int)HRESULT.S_OK == QueryGetDataUnsafe(ref formatetc))
-                {
-                    try
-                    {
-                        innerData.GetData(ref formatetc, out medium);
-                    }
-                    catch
-                    {
-                    }
-                }
-
-                if (medium.unionmember != IntPtr.Zero)
-                {
-                    if (format.Equals(DataFormats.Bitmap))
-                    {
-                        // as/urt 140870 -- GDI+ doesn't own this HBITMAP, but we can't
-                        // delete it while the object is still around.  So we have to do the really expensive
-                        // thing of cloning the image so we can release the HBITMAP.
-
-                        // This bitmap is created by the com object which originally copied the bitmap to tbe
-                        // clipboard. We call Add here, since DeleteObject calls Remove.
-                        Image clipboardImage = Image.FromHbitmap(medium.unionmember);
-                        if (clipboardImage != null)
-                        {
-                            Image firstImage = clipboardImage;
-                            clipboardImage = (Image)clipboardImage.Clone();
-                            Gdi32.DeleteObject(medium.unionmember);
-                            firstImage.Dispose();
-                        }
-                        data = clipboardImage;
-                    }
-                }
-
-                return data;
-            }
-
-            /// <summary>
-            ///  Extracts a managed Object from the innerData of the specified
-            ///  format. This is the base of the OLE to managed conversion.
-            /// </summary>
-            private object GetDataFromBoundOleDataObject(string format, out bool done)
-            {
-                object data = null;
-                done = false;
-                try
-                {
-                    data = GetDataFromOleOther(format);
-                    if (data == null)
-                    {
-                        data = GetDataFromOleHGLOBAL(format, out done);
-                    }
-                    if (data == null && !done)
-                    {
-                        data = GetDataFromOleIStream(format);
-                    }
-                }
-                catch (Exception e)
-                {
-                    Debug.Fail(e.ToString());
-                }
-                return data;
-            }
-
-            /// <summary>
-            ///  Creates an Stream from the data stored in handle.
-            /// </summary>
-            private Stream ReadByteStreamFromHandle(IntPtr handle, out bool isSerializedObject)
-            {
-                IntPtr ptr = Kernel32.GlobalLock(handle);
-                if (ptr == IntPtr.Zero)
-                {
-                    throw new ExternalException(SR.ExternalException, (int)HRESULT.E_OUTOFMEMORY);
-                }
-                try
-                {
-                    int size = Kernel32.GlobalSize(handle);
-                    byte[] bytes = new byte[size];
-                    Marshal.Copy(ptr, bytes, 0, size);
-                    int index = 0;
-
-                    // The object here can either be a stream or a serialized
-                    // object.  We identify a serialized object by writing the
-                    // bytes for the guid serializedObjectID at the front
-                    // of the stream.  Check for that here.
-                    //
-                    if (size > serializedObjectID.Length)
-                    {
-                        isSerializedObject = true;
-                        for (int i = 0; i < serializedObjectID.Length; i++)
-                        {
-                            if (serializedObjectID[i] != bytes[i])
-                            {
-                                isSerializedObject = false;
-                                break;
-                            }
-                        }
-
-                        // Advance the byte pointer.
-                        //
-                        if (isSerializedObject)
-                        {
-                            index = serializedObjectID.Length;
-                        }
-                    }
-                    else
-                    {
-                        isSerializedObject = false;
-                    }
-
-                    return new MemoryStream(bytes, index, bytes.Length - index);
-                }
-                finally
-                {
-                    Kernel32.GlobalUnlock(handle);
-                }
-            }
-
-            /// <summary>
-            ///  Creates a new instance of the Object that has been persisted into the
-            ///  handle.
-            /// </summary>
-            private object ReadObjectFromHandle(IntPtr handle, bool restrictDeserialization)
-            {
-                Stream stream = ReadByteStreamFromHandle(handle, out bool isSerializedObject);
-                if (isSerializedObject)
-                {
-                    return ReadObjectFromHandleDeserializer(stream, restrictDeserialization);
-                }
-
-                return stream;
-            }
-
-            private static object ReadObjectFromHandleDeserializer(Stream stream, bool restrictDeserialization)
-            {
-                BinaryFormatter formatter = new BinaryFormatter();
-                if (restrictDeserialization)
-                {
-                    formatter.Binder = new BitmapBinder();
-                }
-                formatter.AssemblyFormat = FormatterAssemblyStyle.Simple;
-                return formatter.Deserialize(stream);
-            }
-
-            /// <summary>
-            ///  Parses the HDROP format and returns a list of strings using
-            ///  the DragQueryFile function.
-            /// </summary>
-            private string[] ReadFileListFromHandle(IntPtr hdrop)
-            {
-                uint count = Shell32.DragQueryFileW(hdrop, 0xFFFFFFFF, null);
-                if (count == 0)
-                {
-                    return null;
-                }
-
-                var sb = new StringBuilder(Kernel32.MAX_PATH);
-                var files = new string[count];
-                for (uint i = 0; i < count; i++)
-                {
-                    uint charlen = Shell32.DragQueryFileW(hdrop, i, sb);
-                    if (charlen == 0)
-                    {
-                        continue;
-                    }
-
-                    string s = sb.ToString(0, (int)charlen);
-                    string fullPath = Path.GetFullPath(s);
-                    files[i] = s;
-                }
-
-                return files;
-            }
-
-            /// <summary>
-            ///  Creates a string from the data stored in handle. If
-            ///  unicode is set to true, then the string is assume to be Unicode,
-            ///  else DBCS (ASCI) is assumed.
-            /// </summary>
-            private unsafe string ReadStringFromHandle(IntPtr handle, bool unicode)
-            {
-                string stringData = null;
-
-                IntPtr ptr = Kernel32.GlobalLock(handle);
-                try
-                {
-                    if (unicode)
-                    {
-                        stringData = new string((char*)ptr);
-                    }
-                    else
-                    {
-                        stringData = new string((sbyte*)ptr);
-                    }
-                }
-                finally
-                {
-                    Kernel32.GlobalUnlock(handle);
-                }
-
-                return stringData;
-            }
-
-            private unsafe string ReadHtmlFromHandle(IntPtr handle)
-            {
-                IntPtr ptr = Kernel32.GlobalLock(handle);
-                try
-                {
-                    int size = Kernel32.GlobalSize(handle);
-                    return Encoding.UTF8.GetString((byte*)ptr, size - 1);
-                }
-                finally
-                {
-                    Kernel32.GlobalUnlock(handle);
-                }
-            }
-
-            //=------------------------------------------------------------------------=
-            // IDataObject
-            //=------------------------------------------------------------------------=
-            public virtual object GetData(string format, bool autoConvert)
-            {
-                object baseVar = GetDataFromBoundOleDataObject(format, out bool done);
-                object original = baseVar;
-
-                if (!done && autoConvert && (baseVar == null || baseVar is MemoryStream))
-                {
-                    string[] mappedFormats = GetMappedFormats(format);
-                    if (mappedFormats != null)
-                    {
-                        for (int i = 0; ((!done) && (i < mappedFormats.Length)); i++)
-                        {
-                            if (!format.Equals(mappedFormats[i]))
-                            {
-                                baseVar = GetDataFromBoundOleDataObject(mappedFormats[i], out done);
-                                if (!done && baseVar != null && !(baseVar is MemoryStream))
-                                {
-                                    original = null;
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                }
-
-                if (original != null)
-                {
-                    return original;
-                }
-                else
-                {
-                    return baseVar;
-                }
-            }
-
-            public virtual object GetData(string format)
-            {
-                return GetData(format, true);
-            }
-
-            public virtual object GetData(Type format)
-            {
-                return GetData(format.FullName);
-            }
-
-            public virtual void SetData(string format, bool autoConvert, object data)
-            {
-            }
-
-            public virtual void SetData(string format, object data)
-            {
-                SetData(format, true, data);
-            }
-
-            public virtual void SetData(Type format, object data)
-            {
-                SetData(format.FullName, data);
-            }
-
-            public virtual void SetData(object data)
-            {
-                if (data is ISerializable)
-                {
-                    SetData(DataFormats.Serializable, data);
-                }
-                else
-                {
-                    SetData(data.GetType(), data);
-                }
-            }
-
-            private int QueryGetDataUnsafe(ref FORMATETC formatetc)
-            {
-                return innerData.QueryGetData(ref formatetc);
-            }
-
-            private int QueryGetDataInner(ref FORMATETC formatetc)
-            {
-                return innerData.QueryGetData(ref formatetc);
-            }
-
-            public virtual bool GetDataPresent(Type format)
-            {
-                return GetDataPresent(format.FullName);
-            }
-
-            private bool GetDataPresentInner(string format)
-            {
-                Debug.Assert(innerData != null, "You must have an innerData on all DataObjects");
-                FORMATETC formatetc = new FORMATETC
-                {
-                    cfFormat = unchecked((short)(ushort)(DataFormats.GetFormat(format).Id)),
-                    dwAspect = DVASPECT.DVASPECT_CONTENT,
-                    lindex = -1
-                };
-
-                for (int i = 0; i < ALLOWED_TYMEDS.Length; i++)
-                {
-                    formatetc.tymed |= ALLOWED_TYMEDS[i];
-                }
-
-                int hr = QueryGetDataUnsafe(ref formatetc);
-                return hr == (int)HRESULT.S_OK;
-            }
-
-            public virtual bool GetDataPresent(string format, bool autoConvert)
-            {
-                bool baseVar = GetDataPresentInner(format);
-
-                if (!baseVar && autoConvert)
-                {
-                    string[] mappedFormats = GetMappedFormats(format);
-                    if (mappedFormats != null)
-                    {
-                        for (int i = 0; i < mappedFormats.Length; i++)
-                        {
-                            if (!format.Equals(mappedFormats[i]))
-                            {
-                                baseVar = GetDataPresentInner(mappedFormats[i]);
-                                if (baseVar)
-                                {
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                }
-                return baseVar;
-            }
-
-            public virtual bool GetDataPresent(string format)
-            {
-                return GetDataPresent(format, true);
-            }
-
-            public virtual string[] GetFormats(bool autoConvert)
-            {
-                Debug.Assert(innerData != null, "You must have an innerData on all DataObjects");
-
-                IEnumFORMATETC enumFORMATETC = null;
-                ArrayList formats = new ArrayList();
-                try
-                {
-                    enumFORMATETC = innerData.EnumFormatEtc(DATADIR.DATADIR_GET);
-                }
-                catch
-                {
-                }
-
-                if (enumFORMATETC != null)
-                {
-                    enumFORMATETC.Reset();
-
-                    FORMATETC[] formatetc = new FORMATETC[] { new FORMATETC() };
-                    int[] retrieved = new int[] { 1 };
-
-                    while (retrieved[0] > 0)
-                    {
-                        retrieved[0] = 0;
-                        try
-                        {
-                            enumFORMATETC.Next(1, formatetc, retrieved);
-                        }
-                        catch
-                        {
-                        }
-
-                        if (retrieved[0] > 0)
-                        {
-                            string name = DataFormats.GetFormat(formatetc[0].cfFormat).Name;
-                            if (autoConvert)
-                            {
-                                string[] mappedFormats = GetMappedFormats(name);
-                                for (int i = 0; i < mappedFormats.Length; i++)
-                                {
-                                    formats.Add(mappedFormats[i]);
-                                }
-                            }
-                            else
-                            {
-                                formats.Add(name);
-                            }
-                        }
-                    }
-                }
-
-                string[] temp = new string[formats.Count];
-                formats.CopyTo(temp, 0);
-                return GetDistinctStrings(temp);
-            }
-
-            public virtual string[] GetFormats()
-            {
-                return GetFormats(true);
-            }
-        }
-
-        //--------------------------------------------------------------------------
-        // Data Store
-        //--------------------------------------------------------------------------
-
-        private class DataStore : IDataObject
-        {
-            private class DataStoreEntry
-            {
-                public object data;
-                public bool autoConvert;
-
-                public DataStoreEntry(object data, bool autoConvert)
-                {
-                    this.data = data;
-                    this.autoConvert = autoConvert;
-                }
-            }
-
-            private readonly Hashtable data = new Hashtable(BackCompatibleStringComparer.Default);
-
-            public DataStore()
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: Constructed DataStore");
-            }
-
-            public virtual object GetData(string format, bool autoConvert)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: GetData: " + format + ", " + autoConvert.ToString());
-                if (string.IsNullOrWhiteSpace(format))
-                {
-                    return null;
-                }
-
-                DataStoreEntry dse = (DataStoreEntry)data[format];
-                object baseVar = null;
-                if (dse != null)
-                {
-                    baseVar = dse.data;
-                }
-                object original = baseVar;
-
-                if (autoConvert
-                    && (dse == null || dse.autoConvert)
-                    && (baseVar == null || baseVar is MemoryStream))
-                {
-                    string[] mappedFormats = GetMappedFormats(format);
-                    if (mappedFormats != null)
-                    {
-                        for (int i = 0; i < mappedFormats.Length; i++)
-                        {
-                            if (!format.Equals(mappedFormats[i]))
-                            {
-                                DataStoreEntry found = (DataStoreEntry)data[mappedFormats[i]];
-                                if (found != null)
-                                {
-                                    baseVar = found.data;
-                                }
-                                if (baseVar != null && !(baseVar is MemoryStream))
-                                {
-                                    original = null;
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                }
-
-                if (original != null)
-                {
-                    return original;
-                }
-                else
-                {
-                    return baseVar;
-                }
-            }
-
-            public virtual object GetData(string format)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: GetData: " + format);
-                return GetData(format, true);
-            }
-
-            public virtual object GetData(Type format)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: GetData: " + format.FullName);
-                return GetData(format.FullName);
-            }
-
-            public virtual void SetData(string format, bool autoConvert, object data)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: SetData: " + format + ", " + autoConvert.ToString() + ", " + data?.ToString() ?? "(null)");
-                if (string.IsNullOrWhiteSpace(format))
-                {
-                    if (format == null)
-                    {
-                        throw new ArgumentNullException(nameof(format));
-                    }
-
-                    throw new ArgumentException(SR.DataObjectWhitespaceEmptyFormatNotAllowed, nameof(format));
-                }
-
-                // We do not have proper support for Dibs, so if the user explicitly asked
-                // for Dib and provided a Bitmap object we can't convert.  Instead, publish as an HBITMAP
-                // and let the system provide the conversion for us.
-                //
-                if (data is Bitmap && format.Equals(DataFormats.Dib))
-                {
-                    if (autoConvert)
-                    {
-                        format = DataFormats.Bitmap;
-                    }
-                    else
-                    {
-                        throw new NotSupportedException(SR.DataObjectDibNotSupported);
-                    }
-                }
-
-                this.data[format] = new DataStoreEntry(data, autoConvert);
-            }
-            public virtual void SetData(string format, object data)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: SetData: " + format + ", " + data?.ToString() ?? "(null)");
-                SetData(format, true, data);
-            }
-
-            public virtual void SetData(Type format, object data)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: SetData: " + format?.FullName ?? "(null)" + ", " + data?.ToString() ?? "(null)");
-                if (format == null)
-                {
-                    throw new ArgumentNullException(nameof(format));
-                }
-
-                SetData(format.FullName, data);
-            }
-
-            public virtual void SetData(object data)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: SetData: " + data?.ToString() ?? "(null)");
-                if (data == null)
-                {
-                    throw new ArgumentNullException(nameof(data));
-                }
-
-                if (data is ISerializable
-                    && !this.data.ContainsKey(DataFormats.Serializable))
-                {
-                    SetData(DataFormats.Serializable, data);
-                }
-
-                SetData(data.GetType(), data);
-            }
-
-            public virtual bool GetDataPresent(Type format)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: GetDataPresent: " + format.FullName);
-                return GetDataPresent(format.FullName);
-            }
-
-            public virtual bool GetDataPresent(string format, bool autoConvert)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: GetDataPresent: " + format + ", " + autoConvert.ToString());
-                if (string.IsNullOrWhiteSpace(format))
-                {
-                    return false;
-                }
-
-                if (!autoConvert)
-                {
-                    Debug.Assert(data != null, "data must be non-null");
-                    return data.ContainsKey(format);
-                }
-                else
-                {
-                    string[] formats = GetFormats(autoConvert);
-                    Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore:  got " + formats.Length.ToString(CultureInfo.InvariantCulture) + " formats from get formats");
-                    Debug.Assert(formats != null, "Null returned from GetFormats");
-                    for (int i = 0; i < formats.Length; i++)
-                    {
-                        Debug.Assert(formats[i] != null, "Null format inside of formats at index " + i.ToString(CultureInfo.InvariantCulture));
-                        if (format.Equals(formats[i]))
-                        {
-                            Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: GetDataPresent: returning true");
-                            return true;
-                        }
-                    }
-                    Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: GetDataPresent: returning false");
-                    return false;
-                }
-            }
-
-            public virtual bool GetDataPresent(string format)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: GetDataPresent: " + format);
-                return GetDataPresent(format, true);
-            }
-
-            public virtual string[] GetFormats(bool autoConvert)
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: GetFormats: " + autoConvert.ToString());
-                Debug.Assert(data != null, "data collection can't be null");
-                Debug.Assert(data.Keys != null, "data Keys collection can't be null");
-
-                string[] baseVar = new string[data.Keys.Count];
-                data.Keys.CopyTo(baseVar, 0);
-                Debug.Assert(baseVar != null, "Collections should never return NULL arrays!!!");
-                if (autoConvert)
-                {
-                    Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: applying autoConvert");
-
-                    ArrayList formats = new ArrayList();
-                    for (int i = 0; i < baseVar.Length; i++)
-                    {
-                        Debug.Assert(data[baseVar[i]] != null, "Null item in data collection with key '" + baseVar[i] + "'");
-                        if (((DataStoreEntry)data[baseVar[i]]).autoConvert)
-                        {
-                            string[] cur = GetMappedFormats(baseVar[i]);
-                            Debug.Assert(cur != null, "GetMappedFormats returned null for '" + baseVar[i] + "'");
-                            for (int j = 0; j < cur.Length; j++)
-                            {
-                                formats.Add(cur[j]);
-                            }
-                        }
-                        else
-                        {
-                            formats.Add(baseVar[i]);
-                        }
-                    }
-
-                    string[] temp = new string[formats.Count];
-                    formats.CopyTo(temp, 0);
-                    baseVar = GetDistinctStrings(temp);
-                }
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: returing " + baseVar.Length.ToString(CultureInfo.InvariantCulture) + " formats from GetFormats");
-                return baseVar;
-            }
-            public virtual string[] GetFormats()
-            {
-                Debug.WriteLineIf(CompModSwitches.DataObject.TraceVerbose, "DataStore: GetFormats");
-                return GetFormats(true);
-            }
-        }
-
-        /// <summary>
-        ///  Binder that restricts DataObject content deserialization to Bitmap type and
-        ///  serialization to strings and Bitmaps.
-        ///  Deserialization of known safe types(strings and arrays of primitives) does not invoke the binder.
-        /// </summary>
-        private class BitmapBinder : SerializationBinder
-        {
-            // Bitmap type lives in defferent assemblies in the .NET Framework and in .NET Core.
-            // However we allow desktop content to be deserializated in Core and Core content
-            // deserialized on desktop. To support this roundtrip,
-            // Bitmap type identity is unified to the desktop type during serialization
-            // and we use the desktop type name when filtering as well.
-            private static readonly string s_allowedTypeName = "System.Drawing.Bitmap";
-            private static readonly string s_allowedAssemblyName = "System.Drawing";
-            // PublicKeyToken=b03f5f7f11d50a3a
-            private static readonly byte[] s_allowedToken = new byte[] { 0xB0, 0x3F, 0x5F, 0x7F, 0x11, 0xD5, 0x0A, 0x3A };
-
-            /// <summary>
-            ///  Only safe to deserialize types are bypassing this callback, Strings
-            ///  and arrays of primitive types in particular. We are explicitly allowing
-            ///  System.Drawing.Bitmap type to bind using the default binder.
-            /// </summary>
-            /// <param name="assemblyName"></param>
-            /// <param name="typeName"></param>
-            /// <returns>null - continue with the default binder.</returns>
-            public override Type BindToType(string assemblyName, string typeName)
-            {
-                if (string.CompareOrdinal(typeName, s_allowedTypeName) == 0)
-                {
-                    AssemblyName nameToBind = null;
-                    try
-                    {
-                        nameToBind = new AssemblyName(assemblyName);
-                    }
-                    catch
-                    {
-                    }
-                    if (nameToBind != null)
-                    {
-                        if (string.CompareOrdinal(nameToBind.Name, s_allowedAssemblyName) == 0)
-                        {
-                            byte[] tokenToBind = nameToBind.GetPublicKeyToken();
-                            if ((tokenToBind != null) &&
-                                (s_allowedToken != null) &&
-                                (tokenToBind.Length == s_allowedToken.Length))
-                            {
-                                bool block = false;
-                                for (int i = 0; i < s_allowedToken.Length; i++)
-                                {
-                                    if (s_allowedToken[i] != tokenToBind[i])
-                                    {
-                                        block = true;
-                                        break;
-                                    }
-                                }
-                                if (!block)
-                                {
-                                    return null;
-                                }
-                            }
-                        }
-                    }
-                }
-                throw new RestrictedTypeDeserializationException(SR.UnexpectedClipboardType);
-            }
-
-            /// <summary>
-            ///  Bitmap and string types are safe type to serialize/deserialize.
-            /// </summary>
-            /// <param name="serializedType"></param>
-            /// <param name="assemblyName"></param>
-            /// <param name="typeName"></param>
-            public override void BindToName(Type serializedType, out string assemblyName, out string typeName)
-            {
-                // null strings will follow the default codepath in BinaryFormatter
-                assemblyName = null;
-                typeName = null;
-                if (serializedType != null && !serializedType.Equals(typeof(string)) && !serializedType.Equals(typeof(Bitmap)))
-                {
-                    throw new SerializationException(string.Format(SR.UnexpectedTypeForClipboardFormat, serializedType.FullName));
-                }
-            }
-        }
-
-        /// <summary>
-        ///  This exception is used to indicate that clipboard contains a serialized
-        ///  managed object that contains unexpected types and that we should stop processing this data.
-        /// </summary>
-        private class RestrictedTypeDeserializationException : Exception
-        {
-            public RestrictedTypeDeserializationException(string message) : base(message)
-            {
-            }
         }
     }
 }
